# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel OpenApi Documentação - Swagger",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/auth/"
        }
    ],
    "paths": {
...
        "/paciente/store": {
            "post": {
                "tags": [
                    "Paciente"
                ],
                "summary": "Paciente store",
                "operationId": "e220b96ab5da13dda5f90fa84f29966a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "pac_nome",
                                    "pac_dataNascimento"
                                ],
                                "properties": {
                                    "pac_nome": {
                                        "type": "string",
                                        "example": "Fulano Pereira dos Santos"
                                    },
                                    "pac_dataNascimento": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "16/08/1989"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/paciente/update/{id}": {
            "put": {
                "tags": [
                    "Paciente"
                ],
                "summary": "Paciente update",
                "operationId": "ab6127ae7cafe3a8619783e507224557",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "pac_nome": {
                                        "type": "string",
                                        "example": "Fulano Pereira dos Santos"
                                    },
                                    "pac_dataNascimento": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "16/08/1989"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/paciente/delete/{id}": {
            "delete": {
                "tags": [
                    "Paciente"
                ],
                "summary": "Paciente delete",
                "operationId": "a6ac4a6be77b3ed89894ffc1ca821e5f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/paciente/show": {
            "get": {
                "tags": [
                    "Paciente"
                ],
                "summary": "Paciente show",
                "operationId": "5073264e17dc29cea047696ff50d2b21",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/plano-saude/store": {
            "post": {
                "tags": [
                    "Plano de Saúde"
                ],
                "summary": "Plano de Saúde store",
                "operationId": "ed0ffcdaee82047faeebf7f363682418",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "plano_descricao",
                                    "plano_telefone"
                                ],
                                "properties": {
                                    "plano_descricao": {
                                        "type": "string",
                                        "example": "Unimed"
                                    },
                                    "plano_telefone": {
                                        "type": "string",
                                        "example": "(33) 3522-1526"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/plano-saude/update/{id}": {
            "put": {
                "tags": [
                    "Plano de Saúde"
                ],
                "summary": "Plano de Saúde update",
                "operationId": "f8b5f80bfd8527533ee1428562a27018",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "plano_descricao": {
                                        "type": "string",
                                        "example": "Unimed"
                                    },
                                    "plano_telefone": {
                                        "type": "string",
                                        "example": "(33) 3522-1526"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/plano-saude/delete/{id}": {
            "delete": {
                "tags": [
                    "Plano de Saúde"
                ],
                "summary": "Plano de Saúde delete",
                "operationId": "657f1ba9d9ca203fd7761b9eebca9e58",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/plano-saude/show": {
            "get": {
                "tags": [
                    "Plano de Saúde"
                ],
                "summary": "Plano de Saúde show",
                "operationId": "8f6a9b5889490c7ae4117a4e2e2a12f2",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearer_token": {
                "type": "apiKey",
                "description": "Para realizar login adione Bearer + o token. Exemplo: Bearer hjfdsfkhjhjerryue",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}